data/activity/requirements.txt
# Activity Consumer Dependencies
kafka-python==2.0.2
psycopg2-binary==2.9.9
